requirements.txt
openai-whisper
pandas
pydub
tqdm
torch
torchaudio
